Refactor project structure and update dependencies
Removed internal/README.md and restructured internal packages to lib. Enhanced the Earthfile for streamlined development and build processes, updated import paths, and cleaned up go.mod and go.sum files accordingly.
