+++ lib/math/math.go
